+++ lib/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: lcosta-g <lcosta-g@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/09 13:21:31 by lcosta-g          #+#    #+#             */
-/*   Updated: 2025/02/27 15:52:38 by lcosta-g         ###   ########.fr       */
+/*   Updated: 2025/02/28 22:57:18 by lcosta-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ int					ft_strncmp(const char *s1, const char *s2, size_t n);
 char				*ft_strnstr(const char *big, const char *little,
 						size_t len);
 int					ft_atoi(const char *nptr);
-int					ft_atol(const char *nptr);
+long long			ft_atol(const char *nptr);
 char				*ft_strdup(const char *s);
 
 void				*ft_memset(void *s, int c, size_t n);
